+++ src/main.rs
@@ -5,7 +5,7 @@ use run::run_ztcm;
 
 fn main() {
     let cli: Cli = Cli::parse();
-    println!("{cli:#?}");
+    // println!("{cli:#?}");
     let mut _files_paths: Vec<String> = vec![];
     // let args: Vec<String> = env::args().collect();
     run_ztcm(cli)
